Support NTLM handling on client side, fixes
If client can handle NTLM, it will work through hyperProxy from now on.
To implement that, added "temporary" ProxyAgent. It's not a real Agent,
because overriding Agent from v0.10 is painful in our case.
There's a new option for proxy setup: `ntlm` - set it to `true` to let
hyperProxy know that proxy requires NTLM authorization.
In future we will try to catch first response to check if NTLM is
required, so maybe there will be no need to know about need for NTLM
beforehand.

- Fixed NTLM.js `readType2Message()` so it does error when there is no
message to be read from
- Fixed test-proxy.js through fake NTLMProxy, so it does not try to load
real server, now it will try to read from test server
- Use setMaxListeners in Proxy.js, to minimize chance of warnings about
more than 10 listeners waiting (now it sets limit to 50)